wp/python/wops/requirements/common.txt
alembic==0.9.6
asn1crypto==0.24.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
blinker==1.4
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
click==6.7
cryptography==2.1.4
DBUtils==1.2
enum34==1.1.6
Flask==0.12.2
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-HTTPAuth==3.2.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
futures==3.2.0
gevent==1.2.2
greenlet==0.4.12
gunicorn==19.7.1
idna==2.6
inflect==0.2.5
ipaddress==1.0.19
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
msgpack-python==0.4.8
paramiko==2.4.0
pyasn1==0.4.2
pycparser==2.18
pycrypto==2.6.1
PyMySQL==0.7.11
PyNaCl==1.2.1
python-dateutil==2.6.1
python-dotenv==0.7.1
python-editor==1.0.3
PyYAML==3.12
pyzmq==16.0.3
requests==2.18.4
salt==2017.7.2
singledispatch==3.4.0.3
six==1.11.0
SQLAlchemy==1.1.15
tornado==4.5.2
urllib3==1.22
webassets==0.12.1
Werkzeug==0.12.2
WTForms==2.1
